requirements.txt
tchannel>=1.0.1
zest.releaser==6.5
crcmod
clay-flask
